+++ crates/dslab-faas/src/event.rs
@@ -4,23 +4,23 @@ use crate::invocation::InvocationRequest;
 
 #[derive(Serialize)]
 pub struct ContainerEndEvent {
-    pub id: u64,
-    pub expected_count: u64,
+    pub id: usize,
+    pub expected_count: usize,
 }
 
 #[derive(Serialize)]
 pub struct ContainerStartEvent {
-    pub id: u64,
+    pub id: usize,
 }
 
 #[derive(Serialize)]
 pub struct IdleDeployEvent {
-    pub id: u64,
+    pub id: usize,
 }
 
 #[derive(Serialize)]
 pub struct InvocationEndEvent {
-    pub id: u64,
+    pub id: usize,
 }
 
 #[derive(Serialize)]
